Taking previous time

+++ src/Services/Alerts/API/Models/StepTimeAsDate.cs
@@ -5,6 +5,6 @@
         public string StepName { get; set; }
         public DateTime StepStart { get; set; }
         public DateTime StepStop { get; set; }
-        public long StepDuration { get; set; }
+        public double StepDuration { get; set; }
     }
 }
